backend/requirements.txt
numpy==1.19.4
Flask==2.0.1
flask-sqlalchemy==2.5.1
pymysql==1.0.2
cryptography==3.4.7
flask-jwt-extended==4.2.1
flask-cors==3.0.10
protobuf==3.17.3
google-auth==1.32.0
python-magic-bin==0.4.14
pyyaml==5.4.1
tabulate==0.8.9
scipy==1.5.4
editdistance==0.5.3
h5py==3.1.0
intervaltree==3.1.0
gdown==3.13.0
scikit-learn==0.24.0
python-dotenv==0.18.0
